requirements.txt
tensorflow==2.16.1
pillow
flask
flask-cors
numpy
werkzeug